cambio2

+++ src/Contraseña.java
@@ -2,37 +2,37 @@ package src;
 
 public class Contraseña {
 
-    public double contraseña1;
-    public double contraseña2;
-    public double contraseña3;
+    public int contraseña1;
+    public int contraseña2;
+    public int contraseña3;
 
-    public Contraseña(double contraseña1, double contraseña2, double contraseña3) {
+    public Contraseña(int contraseña1, int contraseña2, int contraseña3) {
         this.contraseña1 = contraseña1;
         this.contraseña2 = contraseña2;
         this.contraseña3 = contraseña3;
     }
 
-    public double getContraseña1() {
+    public int getContraseña1() {
         return contraseña1;
     }
 
-    public void setContraseña1(double contraseña1) {
+    public void setContraseña1(int contraseña1) {
         this.contraseña1 = contraseña1;
     }
 
-    public double getContraseña2() {
+    public int getContraseña2() {
         return contraseña2;
     }
 
-    public void setContraseña2(double contraseña2) {
+    public void setContraseña2(int contraseña2) {
         this.contraseña2 = contraseña2;
     }
 
-    public double getContraseña3() {
+    public int getContraseña3() {
         return contraseña3;
     }
 
-    public void setContraseña3(double contraseña3) {
+    public void setContraseña3(int contraseña3) {
         this.contraseña3 = contraseña3;
     }
 }
